requirements.txt
streamlit
pandas
numpy
plotly
python-pptx
reportlab
openpyxl
kaleido
matplotlib
